UHF-10349: Move convert functions to their own helper class
