+++ public/modules/custom/grants_handler/src/Plugin/WebformHandler/GrantsHandler.php
@@ -258,51 +258,6 @@ final class GrantsHandler extends WebformHandlerBase {
   }
 
   /**
-   * Convert EUR format value to float.
-   *
-   * @param string|null $value
-   *   Value to be converted.
-   *
-   * @return float|null
-   *   Floated value.
-   */
-  public static function convertToFloat(?string $value = ''): ?float {
-    if (is_null($value)) {
-      return NULL;
-    }
-
-    if ($value === '') {
-      return NULL;
-    }
-
-    $value = str_replace(['€', ',', ' '], ['', '.', ''], $value);
-    return (float) $value;
-  }
-
-  /**
-   * Convert EUR format value to "int" .
-   *
-   * @param string|null $value
-   *   Value to be converted.
-   *
-   * @return int|null
-   *   Int value.
-   */
-  public static function convertToInt(?string $value = ''): ?int {
-    if (is_null($value)) {
-      return NULL;
-    }
-
-    if ($value === '') {
-      return NULL;
-    }
-
-    $value = str_replace(['€', ',', ' ', '_'], ['', '.', '', ''], $value);
-    $value = (int) $value;
-    return $value;
-  }
-
-  /**
    * {@inheritdoc}
    */
   public function defaultConfiguration(): array {
